+++ class/agmono/Mono/NativeMethods.cs
@@ -550,7 +550,7 @@ namespace Mono {
 		public extern static void downloader_abort (IntPtr handle);
 		
 		[DllImport("moon")]
-		public extern static IntPtr downloader_get_response_text (IntPtr handle, string partname, out uint size);
+		public extern static IntPtr downloader_get_response_text (IntPtr handle, string partname, out long size);
 
 		[DllImport("moon")]
 		public extern static void downloader_open (IntPtr handle, string verb, string uri, bool async);
